Cleanup in code, structure etc
