+++ screens/chatOverviews/chatOverviewScreen.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import ChatOverviewList from '../../components/chatOverviewList/chatOverviewComponent'
+
+export default chatOverviews = ({ route, navigation }) => {
+  return (
+    <>
+      <ChatOverviewList route={route} navigation={navigation} />
+    </>
+  )
+}
