+++ screens/chatOverviews/index.js
@@ -1,10 +1,1 @@
-import React from 'react'
-import ChatOverviewList from '../../components/chatOverviewList'
-
-export default chatOverviews = ({ route, navigation }) => {
-  return (
-    <>
-      <ChatOverviewList route={route} navigation={navigation} />
-    </>
-  )
-}
+export { default } from './chatOverviewScreen'
