+++ screens/chatWindowScreen/chatWindowScreen.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+import ChatWindow from '../../components/chatWindow/chatWindowMainComponent'
+
+export default ChatWindowScreen = ({ route, navigation }) => {
+  return <ChatWindow route={route} navigation={navigation} />
+}
