+++ screens/chatWindowScreen/index.js
@@ -1,7 +1,1 @@
-import React from 'react'
-
-import ChatWindow from '../../components/chatWindow'
-
-export default ChatWindowScreen = ({ route, navigation }) => {
-  return <ChatWindow route={route} navigation={navigation} />
-}
+export { default } from './chatWindowScreen'
